구매 파트 합치기 완료

+++ backend/src/main/java/com/farmday/orders/OrdersController.java
@@ -3,6 +3,7 @@ package com.farmday.orders;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -10,10 +11,15 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.farmday.cart.CartDTO;
 import com.farmday.cart.CartService;
+import com.farmday.coupon.service.MembershipCouponServiceImpl;
+import com.farmday.groupdeal.dto.GroupDealDetailResponseDto;
+import com.farmday.groupdeal.dto.GroupDealImageDto;
+import com.farmday.groupdeal.service.GroupDealService;
 import com.farmday.mypage.MembershipGradeDTO;
 import com.farmday.mypage.MypageService;
 import com.farmday.mypage.OrdersCanceledDTO;
 
+import java.util.ArrayList;
 import java.util.List;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -30,6 +36,15 @@ public class OrdersController {
 
     @Autowired
     MypageService mypageService;
+
+    @Autowired
+    GroupDealService groupDealService;
+
+    @Autowired
+    com.farmday.groupdeal.mapper.GroupDealMapper groupDealMapper;
+
+    @Autowired
+    MembershipCouponServiceImpl membershipCouponServiceImpl;
 
     // public OrdersController(OrdersService ordersService, CartService cartService, MypageService mypageService) { 
     //     this.ordersService = ordersService; 
@@ -49,6 +64,7 @@ public class OrdersController {
 
 
     // 구매 시 처리
+    @Transactional(rollbackFor = Exception.class)
     @PostMapping(value = "/orders/insertOrders/{user_id}", consumes = "application/json")
     public ResponseEntity<Void> insertCart(
             @PathVariable("user_id") String user_id,
@@ -57,6 +73,7 @@ public class OrdersController {
         if (user_id == null || user_id.isEmpty() || item == null) {
             return ResponseEntity.badRequest().build();
         }
+        int user_no = ordersService.findUserInfoForOrder(user_id).getUser_no().intValue();
 
         int used_points = item.getUsed_points();
         int current_points = ordersService.findUserPoints(user_id);
@@ -70,6 +87,52 @@ public class OrdersController {
         item.setProduct_total_amount(item.getSubtotal());
 
         String toss_orderid = item.getToss_orderid();
+
+
+
+
+        if (item.getOrder_type().length()>=9 && item.getOrder_type().substring(0, 9).equals("groupdeal")) {
+            String[] splitInfo = item.getOrder_type().split("_");
+            long groupDealId = Long.parseLong(splitInfo[1]);
+            item.setOrder_type("groupdeal")
+            ;
+            ordersService.insertOrders(item);
+
+            // 토스페이먼츠 결제로 생성된 고유 주문번호를 통해 방금 생성된 order의 PK값 찾아오기
+            int order_id = ordersService.findOrdersByTossOrderId(toss_orderid).getOrder_id();
+
+            GroupDealDetailResponseDto groupDealDTO = groupDealService.getGroupDealDetail(groupDealId);
+
+            OrdersItemDTO ordersItem = new OrdersItemDTO();
+            ordersItem.setOrder_id(order_id);
+            ordersItem.setProduct_id(0);
+            ordersItem.setProduct_name(groupDealDTO.getTitle());
+            ordersItem.setGroup_deal_id(groupDealDTO.getGroupDealId().intValue());
+            ordersItem.setPrice_at_order(groupDealDTO.getDealPrice().intValue());
+            ordersItem.setQuantity(Integer.parseInt(splitInfo[2]));
+            ordersItem.setLine_total_amount(groupDealDTO.getDealPrice().intValue()*Integer.parseInt(splitInfo[2]));
+
+            ordersService.insertOrdersItem(ordersItem);
+
+            groupDealService.joinGroupDeal(user_id, groupDealId, Integer.parseInt(splitInfo[2]));
+
+
+            int order_item_id = ordersService.findOrderItemIdOfGroupDealOrder(order_id);
+
+            DeliveryDTO delivery = new DeliveryDTO();
+            delivery.setOrder_item_id(order_item_id);
+            delivery.setDelivery_status("배송준비");
+
+            ordersService.insertOrdersItemIntoDelivery(delivery);
+
+            membershipCouponServiceImpl.deleteUsedCoupon(user_no, item.getCouponId());
+
+
+            System.out.println("그룹딜 주문 처리 완료");
+
+            return ResponseEntity.noContent().build();
+        }
+
         ordersService.insertOrders(item);
 
 
@@ -84,13 +147,21 @@ public class OrdersController {
             ordersItem.setOrder_id(order_id);
             ordersItem.setProduct_id(cart.getProduct_id());
             ordersItem.setProduct_name(cart.getProduct_name());
-            ordersItem.setDiscount_amount(0);
+            ordersItem.setGroup_deal_id(0);
             ordersItem.setPrice_at_order(cart.getPrice());
             ordersItem.setQuantity(cart.getQuantity());
             ordersItem.setLine_total_amount(cart.getPrice()*cart.getQuantity());
 
+            // 트랜잭션을 통한 Database 무결성 보장
+            boolean ok = ordersService.updateProductAmount(cart.getProduct_id(), -cart.getQuantity());
+
+            if (!ok) {
+                throw new IllegalStateException("재고 부족으로 주문을 진행할 수 없습니다.");
+            }
+
             ordersService.insertOrdersItem(ordersItem);
             cartService.deleteCart(cart.getCart_id());
+
         }
 
         List<OrdersItemDTO> ordersItemList = ordersService.findOrdersItemByOrderId(order_id);
@@ -103,7 +174,9 @@ public class OrdersController {
             ordersService.insertOrdersItemIntoDelivery(delivery);
         }
         
-        
+        int deleted = membershipCouponServiceImpl.deleteUsedCoupon(user_no, item.getCouponId());
+
+        System.out.println("deleted : "+deleted+", user_no : "+user_no+", coupon_id : "+item.getCouponId());
 
         // return ResponseEntity.status(HttpStatus.CREATED).build();
         return ResponseEntity.noContent().build();
@@ -125,12 +198,17 @@ public class OrdersController {
                 continue;
             }
             int order_amount = ordersItem.size();
+            dto.setOrder_amount(order_amount);
+
+            List<OrdersImgListDTO> imgList = ordersService.findOrdersImgListByOrderId(dto.getOrder_id());
+            
+            dto.setOrders_img_list(imgList);
+
             int product_id = ordersItem.get(0).getProduct_id();
             String thumbnail_name = ordersItem.get(0).getProduct_name();
             String thumbnail_img = mypageService.findProductMainImage(product_id);
 
 
-            dto.setOrder_amount(order_amount);
             dto.setThumbnail_name(thumbnail_name);
             dto.setThumbnail_img(thumbnail_img);
             dto.setThumbnail_id(product_id);
@@ -139,10 +217,51 @@ public class OrdersController {
         return userOrders;
     }
 
+
+    @GetMapping("/orders/findAllGroupDealOrdersByUserId")
+    public List<OrdersDTO> findAllGroupDealOrdersByUserId(@RequestParam("user_id")String user_id) throws Exception {
+
+        List<OrdersDTO> userOrders = ordersService.findAllGroupDealOrdersByUserId(user_id);
+
+        for (OrdersDTO dto : userOrders) {
+
+            OrdersItemDTO ordersItem = ordersService.findGroupDealItemByOrderId(dto.getOrder_id());
+            if (ordersItem == null) {
+                continue;
+            }
+
+            List<GroupDealImageDto> imageList = groupDealMapper.selectGroupDealImages((long)ordersItem.getGroup_deal_id());
+
+            List<OrdersImgListDTO> ordersImgList = new ArrayList<>();
+
+            for (GroupDealImageDto imgDto : imageList) {
+                OrdersImgListDTO ordersImg = new OrdersImgListDTO();
+                ordersImg.setProduct_id(ordersItem.getGroup_deal_id());
+                ordersImg.setMain_image(imgDto.getImageUrl());
+                ordersImgList.add(ordersImg);
+            }
+
+            dto.setOrders_img_list(ordersImgList);
+
+
+        }
+
+        return userOrders;
+    }
+
+
+
     @GetMapping("/orders/findOrdersItemByOrderId")
     public List<OrdersItemDTO> findOrdersItemByOrderId(@RequestParam("order_id")int order_id) throws Exception {
         return ordersService.findOrdersItemByOrderId(order_id);
     }
+
+    @GetMapping("/orders/findGroupDealItemByOrderId")
+    public OrdersItemDTO findGroupDealItemByOrderId(@RequestParam("order_id")int order_id) throws Exception {
+        return ordersService.findGroupDealItemByOrderId(order_id);
+    }
+
+
 
     @GetMapping("/orders/findOrdersItemById")
     public OrdersItemDTO findOrdersItemById(@RequestParam("order_item_id")int order_item_id) throws Exception {
@@ -218,6 +337,8 @@ public class OrdersController {
 
         ordersService.insertOrdersItemIntoCancel(cancel);
 
+        ordersService.updateProductAmount(ordersItem.getProduct_id(),ordersItem.getQuantity());
+
         // 취소 및 환불 완료 코드인 R1으로 변경
         ordersService.changeOrdersItemStatus(order_item_id, "R1");
 
@@ -287,7 +408,5 @@ public class OrdersController {
 
         return ResponseEntity.ok("");
     }
-
-
 
 }
